requirements.txt
replicate>=0.22.0
gradio>=4.19.2
python-dotenv>=1.0.0 